+++ src/graph/geometry_impl.h
@@ -115,7 +115,7 @@ line_distance(coordinate_t __source, coordinate_t __destination, coordinate_t __
  * @return
  */
 inline coordinate_t
-interpolate_linear(coordinate_t source, coordinate_t destination, float relative) {
+interpolate_linear(coordinate_t source, coordinate_t destination, double relative) {
     destination -= source;
     destination *= relative;
     return source + destination;
